admin controller

+++ eX-Portal/eX-Portal/Controllers/AdminController.cs
@@ -16,7 +16,8 @@ namespace eX_Portal.Controllers {
     }//function index()
 
     public ActionResult Account() {
-      ViewBag.Title = "Accounts";
+            if (!exLogic.User.hasAccess("ACCOUNT.VIEW")) return RedirectToAction("NoAccess", "Home");
+            ViewBag.Title = "Accounts";
       String SQL = "SELECT \n" +
         "  [AccountId],\n" +
         "  [Name],\n" +
@@ -29,8 +30,8 @@ namespace eX_Portal.Controllers {
         "  [MSTR_Account]";
 
       qView nView = new qView(SQL);
-      if (exLogic.User.hasAccess("ACCOUNT.EDIT")) nView.addMenu("Edit", Url.Action("Edit", new { ID = "_PKey" }));
-      if (exLogic.User.hasAccess("ACCOUNT.DELETE")) nView.addMenu("Delete", Url.Action("Delete", new { ID = "_PKey" }));
+      if (exLogic.User.hasAccess("ACCOUNT.EDIT")) nView.addMenu("Edit", Url.Action("AccountEdit", new { ID = "_PKey" }));
+      if (exLogic.User.hasAccess("ACCOUNT.DELETE")) nView.addMenu("Delete", Url.Action("AccountDelete", new { ID = "_PKey" }));
 
       if (Request.IsAjaxRequest()) {
         Response.ContentType = "text/javascript";
@@ -41,46 +42,119 @@ namespace eX_Portal.Controllers {
     }//Account()
 
     public ActionResult AccountCreate() {
-      ViewBag.Title = "Create Account";
-      return View();
+
+            if (!exLogic.User.hasAccess("ACCOUNT.CREATE")) return RedirectToAction("NoAccess", "Home");
+            ViewBag.Title = "Create Account";
+            var AccountModel = new ViewModel.AccountViewModel
+            {
+                Account = new MSTR_Account(),
+                CountryList = Util.GetCountryLists("Country", "DroneName", "Code", "usp_Portal_DroneServiceType"),
+               
+
+            };
+
+
+            return View(AccountModel);
     }
 
     [HttpPost]
     public ActionResult AccountCreate(MSTR_Account Account) {
-      if (ModelState.IsValid) {
-        Account.IsActive = true;
-        Account.CreatedBy = Util.toInt(Session["UserID"].ToString());
+            if (!exLogic.User.hasAccess("ACCOUNT.CREATE")) return RedirectToAction("NoAccess", "Home");
+            if (ModelState.IsValid) {
+
+                if (Session["UserId"] == null)
+                {
+                    Session["UserId"] = -1;
+                }
+                Account.CreatedBy = Util.toInt(Session["UserID"].ToString());
         Account.CreatedOn = DateTime.Now;
         db.MSTR_Account.Add(Account);
         db.SaveChanges();
+                //setting up binary code after the  account id creation  and saving
+                String CodeBinary = Util.DecToBin(Account.AccountId);
+                String SQL = "Update Mstr_Account set BinaryCode='" + CodeBinary + "' where AccountId=" + Account.AccountId;
+                Util.doSQL(SQL);
         return RedirectToAction("AccountDetail", new { id = Account.AccountId});
       }
-      return View(Account);
+
+            var viewModel = new ViewModel.AccountViewModel
+            {
+                Account = new MSTR_Account(),
+                CountryList = Util.GetCountryLists("Country", "CountryName", "Code", "sp"),
+
+
+
+            };
+            return View(viewModel);
+          
     }
 
-    public ActionResult AccountEdit() {
+    public ActionResult AccountEdit(int  id) {
+      
       ViewBag.Title = "Edit Account";
-      return View();
+            if (!exLogic.User.hasAccess("ACCOUNT.EDIT")) return RedirectToAction("NoAccess", "Home");
+
+            var viewModel = new ViewModel.AccountViewModel
+            {
+
+
+
+                Account = db.MSTR_Account.Find(id),
+
+                CountryList = Util.GetCountryLists("Country", "CountryName", "Code", "sp"),
+            };
+            return View(viewModel);
+           
     }
 
     [HttpPost]
     public ActionResult AccountEdit(Models.MSTR_Account Account ) {
       ViewBag.Title = "Edit Account";
-      try {
+            if (!exLogic.User.hasAccess("ACCOUNT.EDIT")) return RedirectToAction("NoAccess", "Home");
+
+            try
+            {
         if (ModelState.IsValid) {
-          Account.ModifiedBy = Util.toInt(Session["UserID"].ToString());
+
+                    if (Session["UserId"] == null)
+                    {
+                        Session["UserId"] = -1;
+                    }
+                    Account.ModifiedBy = Util.toInt(Session["UserID"].ToString());
           Account.ModifiedOn = DateTime.Now;
-          db.Entry(Account).Property(x => x.CreatedBy).IsModified = false;
-          db.Entry(Account).Property(x => x.CreatedOn).IsModified = false;
-          db.MSTR_Account.Add(Account);
-          db.SaveChanges();
-          return RedirectToAction("AccountDetail");
+
+                string    SQL = "Update Mstr_Account set Name='" + Account.Name + "',Code='"
+                                   + Account.Code + "',EmailId='" + Account.EmailId + "',MobileNo='"
+                                   + Account.MobileNo + "',OfficeNo='" + Account.OfficeNo + "',IsActive='" + Account.IsActive + "',ModifiedBy=" + Session["UserId"] +
+                                   ",ModifiedOn='"+ DateTime.Now.ToString("yyyy - MM - dd")+ "', AccountDescription='"+ Account.AccountDescription +
+                                   "',Address1='"+Account.Address1  +"', Address2='"+Account.Address2 + "' ,Address3='"+ Account.Address3 + 
+                                   "',CountryCode="+ Account.CountryCode +" where AccountId=" + Account.AccountId;
+
+                    int id = Util.doSQL(SQL);
+
+                    /* db.Entry(Account).Property(x => x.CreatedBy).IsModified = false;
+                       db.Entry(Account).Property(x => x.CreatedOn).IsModified = false;
+                       db.MSTR_Account.Add(Account);
+                       db.SaveChanges();*/
+                    return RedirectToAction("Account");
         }
       } catch (Exception  ex) {
         //Log the error (uncomment dex variable name and add a line here to write a log.
         return View("InternalError", ex);
       }
-      return View(Account);
+
+            //for the server side validation
+            var viewModel = new ViewModel.AccountViewModel
+            {
+
+
+
+                Account = db.MSTR_Account.Find(Account.AccountId),
+
+                CountryList = Util.GetCountryLists("Country", "CountryName", "Code", "sp"),
+            };
+            return View(viewModel);
+          
     }//ActionEdit()
 
     public ActionResult AccountDetail([Bind(Prefix = "ID")] int AccountID) {
@@ -99,25 +173,25 @@ namespace eX_Portal.Controllers {
 
 
 
-        public String Delete([Bind(Prefix = "ID")]int AccountID = 0)
+        public String AccountDelete([Bind(Prefix = "ID")]int AccountID = 0)
         {
             if (!exLogic.User.hasAccess("ACCOUNT.DELETE"))
 
                 return Util.jsonStat("ERROR", "Access Denied");
             String SQL = "";
-            Response.ContentType = "text/json";
-            // if (!exLogic.User.hasAccess("USER.DELETE"))
-            //   return Util.jsonStat("ERROR", "Access Denied");
+            Response.ContentType = "text/json";            
 
             //Delete the Account from database if there is no Drone createdby
-            SQL = "SELECT Count(*) FROM MSTR_Drone where OwnerId = " + AccountID;
+            SQL = "SELECT Count(*) FROM MSTR_Drone where AccountId = " + AccountID;
 
             if (Util.getDBInt(SQL) != 0)
-                return Util.jsonStat("ERROR", "You can not delete a the User Attached to another user");
+                return Util.jsonStat("ERROR", "You can not delete a the User Attached Drone");
 
-          
+            SQL= "select Count(*) from MSTR_Parts where SupplierId =" + AccountID;
+            if (Util.getDBInt(SQL) != 0)
+                return Util.jsonStat("ERROR", "You can not delete a the User Attached to Parts");
 
-            SQL = "DELETE FROM [MSTR_Account] WHERE   = " + AccountID;
+            SQL = "DELETE FROM [MSTR_Account] WHERE  AccountId = " + AccountID;
             Util.doSQL(SQL);
 
             return Util.jsonStat("OK");
